pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "petthecord"
description = "Web API for petting any user you want"
readme = "README.md"
version = "v1.0.0"
dependencies = [
    "discord.py==2.4.0",
    "aiohttp==3.10.5",
    "pet-pet-gif@git+https://github.com/camprevail/pet-pet-gif@28a4a5d18b312bcb84e6dcd461c0e16fc6d2bf16",
]
authors = [{name = "Nakidai", email = "plaza521@inbox.ru"}]
classifiers = [
    "License :: OSI Approved :: BSD License",
    "Programming Language :: Python :: 3.12",
]

[project.scripts]
petthecord = "petthecord:main"
